% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjDR.R
\name{adjDR}
\alias{adjDR}
\title{Doubly Robust Cumulative Incidence Curves}
\usage{
adjDR(formula, strata, ref = NULL, data = NULL, times = NULL)
}
\arguments{
\item{formula}{Object containing a `Surv()` object as the response, and the strata and covariates as the predictors.}

\item{strata}{A character string used to identify the strata within the `formula` object.}

\item{ref}{A character input to specify a reference group from `strata`. If this argument is `NULL`, which is by default, the entire sample is used as a reference instead.}

\item{data}{An optional `data.frame` used to interpret the variables described by `formula` argument.}

\item{times}{A series of user-defined time points for which the cumulative incidence should be computed per strata. If `NULL`, which is by default, the unique observation times from the data are used instead.}
}
\value{
A `data.frame` containing the covariate adjusted cumulative incidence curves per strata
}
\description{
Generates covariate adjusted cumulative incidence curves via doubly robust.
}
\examples{
df <- simCSH(N = 1000, k = 2,
  theta_coef = c(-1, -0.5),
  beta_coef = list(c(1, -1, 0.5), c(-1, 1, -0.5)),
  base_haz = c(2, -2))
df_DR <- adjDR(Surv(obs_time, status) ~ treatment + x1 + x2, data = df,
  strata = "treatment", times = seq(0, 1.5, 0.05))
}
