% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjIPW.R
\name{adjIPW}
\alias{adjIPW}
\title{Generate Inverse Probability Weights}
\usage{
adjIPW(formula, data = NULL)
}
\arguments{
\item{formula}{A `formula` object in the form of `y ~ x`, where `x` represents the collection of covariates which should be balanced between the levels of `y`.}

\item{data}{An optional `data.frame` used to interpret the variables described by `formula` argument.}
}
\value{
A `vector` containing the inverse probability weights for each observation
}
\description{
A simple function to generates inverse probability weights via standard logistic regression
}
\examples{
df <- simCSH(N = 1000, k = 2,
  theta_coef = c(-1, -0.5),
  beta_coef = list(c(1, -1, 0.5), c(-1, 1, -0.5)),
  base_haz = c(2, -2))
df$ipw <- adjIPW(cohort ~ age + mutation, data = df)
df_IPW <- survfit(Surv(obs_time, status) ~ cohort, data = df, weights = ipw)
}
