% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPW.R
\name{IPW}
\alias{IPW}
\title{Generate Inverse Probability Weights}
\usage{
IPW(formula, data = NULL)
}
\arguments{
\item{formula}{A `formula` object in the form of `y ~ x`, where `x` represents the collection of covariates which should be balanced between the levels of `y`.}

\item{data}{An optional `data.frame` used to interpret the variables described by `formula` argument.}
}
\value{
A `vector` containing the inverse probability weights for each observation
}
\description{
A simple function to generates inverse probability weights via standard logistic regression
}
\examples{
bc_data <- simCSH(N = 1000, k = 2,
    base_haz = c(1, 1), beta_coef = list(c(-0.5, 2, 0.5), c(0.3, 1, -0.4)))
bc_data$ipw <- IPW(cohort ~ age + mutation, data = bc_data)
library(survival)
bc_cuminc <- survfit(Surv(obs_time, status) ~ cohort, data = bc_data, weights = ipw)
}
