% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longCSH.R
\name{longCSH}
\alias{longCSH}
\title{Extract cumulative incidence curves from a survfit object}
\usage{
longCSH(object, stack = F)
}
\arguments{
\item{object}{a survfit object from which one would like to extract the cumulative incidence estimator.}

\item{strata}{a boolean operator (TRUE / FALSE), which determines whether to return the marginal probabilities per event. If TRUE, which is by default, the cause-specific probabilities are returned as stacked probabilities for subsequent events.}
}
\value{
A `data.frame` with the output from survfit in long format
}
\description{
This helper function can be used to quickly extract the strata, times, and cause-specific probabilities from a survfit object for each event in a long format, such that it can be easily plotted (eg. by ggplot).
}
\examples{
df <- simCSH(N = 1000, k = 2,
  theta_coef = c(-1, -0.5),
  beta_coef = list(c(1, -1, 0.5), c(-1, 1, -0.5)),
  base_haz = c(2, -2))
df_long <- longCSH(survfit(Surv(obs_time, status) ~ treatment, data = df))
}
