% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueCSH.R
\name{trueCSH}
\alias{trueCSH}
\title{Hypothetical true competing risk curve computation}
\usage{
trueCSH(time, k, theta_coef, beta_coef, base_haz)
}
\arguments{
\item{time}{a vector of time points.}

\item{k}{the number of events.}

\item{theta_coef}{vector where each element represents the coefficient of treatment on the hazard of each event.}

\item{beta_coef}{a list where each listed element is a vector of covariate coefficients associated with event k. The first two elements introduce a discrete covariate effect on the outcome, and the third element a continuous covariate effect on the outcome.}

\item{base_haz}{a vector of length 2 with a baseline hazard conditional on the continuous covariate. If these two values differ, scenario 3 is induced, resulting in a non-linear relationship between covariates and outcome.}
}
\value{
A `data.frame` with the hypothetical true curve per treatment condition
}
\description{
This function was developed to calculate the hypothetical true curve for a series of time points, based on the cause-specific hazard model. The hypothetical true curve can be controlled by a list of baseline hazard, and treatment covariate effect coefficients per event. Non-linearity conditional on the continuous covariate can be introduced by choosing two different values for `base_haz`.
}
\examples{
df_true <- trueCSH(time = times, k = 2,
  theta_coef = c(-1, -0.5),
  beta_coef = list(c(1,  -1, 0.5), c(-1, 1, -0.5)),
  base_haz = c(2, -2))
}
